Corregir tipo de variable en string_length y mejorar la documentación en strfuncs.h

+++ lab04/lab04/ej4/a-strfuncs/strfuncs.c
@@ -5,7 +5,7 @@
 #include "strfuncs.h"
 
 size_t string_length(const char *str) {
-    unsigned long length = 0;
+    size_t length = 0;
     while(str[length] != '\0') {
         length++;
     }
